fixing: api product-cart-auth-address

+++ database/migrations/2024_05_21_040652_create_products_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->longText('image3')->nullable();
             $table->integer('size');
             $table->integer('price');
-            $table->float('discount');
+            $table->integer('discount');
             $table->enum('status', ['available', 'sold out'])->default('available');            
             $table->timestamps();
         });
